+++ src/demo/iframed-headless.html
@@ -4,160 +4,12 @@
     </head>
     <body>
         <script type="text/javascript" src="../recorder.js"></script>
-        <talk-recorder id="recorder" bitRate="64k"></talk-recorder>
+        <talk-recorder id="recorder" bitRate="64k" role="iframed"></talk-recorder>
         <script type="text/javascript">
             window.addEventListener("message", (e) => {
                 console.log("iframed received", { e, referrer: document.referrer });
                 parent.postMessage({ type: "iframed.hello" }, e.origin);
             });
-            function getRecorder(type) {
-                switch (type) {
-                    case "new":
-                        return new TalkRecorder();
-                    case "get":
-                        return document.getElementById("recorder");
-                    case "create":
-                        return document.createElement("talk-recorder");
-                }
-            }
-
-            // Record using promise
-            function recordPromise(recorder, type) {
-                return recorder
-                    .record({ type })
-                    .then((audioBlob) => {
-                        console.log("got recording", audioBlob);
-                        return audioBlob;
-                    })
-                    .catch((err) => console.error(err));
-            }
-
-            function PCB(promise, cb) {
-                return promise.then((result) => cb(null, result)).catch((err) => cb(err, null));
-            }
-
-            // Record using callback
-            function recordCB(recorder, type, cb) {
-                PCB(recorder.record({ type }), (err, audioBlob) => {
-                    if (err) {
-                        console.error(err);
-                    } else {
-                        console.log("got recording", audioBlob);
-                    }
-                    cb(err, audioBlob);
-                });
-            }
-
-            // Record using events
-            function recordEvents(recorder, type) {
-                recorder.addEventListener(
-                    "record",
-                    (e) => {
-                        console.log("recording started");
-                    },
-                    { once: true }
-                );
-                recorder.addEventListener(
-                    "stream",
-                    (e) => {
-                        const stream = e.detail.stream;
-                        console.log("recording stream", stream);
-                    },
-                    { once: true }
-                );
-                recorder.addEventListener(
-                    "recorded",
-                    (e) => {
-                        const blob = e.detail.blob;
-                        console.log("recorded blob", blob);
-                    },
-                    { once: true }
-                );
-                recorder.addEventListener(
-                    "error",
-                    (e) => {
-                        const error = e.detail.error;
-                        console.error("recording failed", error);
-                    },
-                    { once: true }
-                );
-                recorder.record({ type });
-            }
-
-            // Displays an audio blob as an audio element.
-            function addRecording(audioBlob, label) {
-                const blobURL = window.URL.createObjectURL(audioBlob);
-                const audioEl = new Audio(blobURL);
-                audioEl.controls = true;
-                Object.assign(audioEl.style, {
-                    display: "inline-block",
-                    height: "32px",
-                    margin: "5px",
-                });
-
-                const recordingEl = document.createElement("div");
-                Object.assign(recordingEl.style, {
-                    display: "flex",
-                    alignItems: "center",
-                });
-
-                recordingEl.appendChild(audioEl);
-                const labelEl = document.createElement("span");
-                labelEl.textContent = ` ${label}`;
-                recordingEl.appendChild(labelEl);
-
-                document.getElementById("recordings").appendChild(recordingEl);
-            }
-
-            // Gets or creates an instance of TalkRecorder using:
-            // 'new' - new TalkRecorder()
-            // 'get' - from an existing '<talk-recorder>' element.
-            // 'create' - from document.createElement('talk-recorder');
-            const recorder = getRecorder("get");
-
-            // Shows how to record an Opus blob.
-            //
-            function recordOpus() {
-                recordPromise(recorder, "opus").then((audioBlob) => addRecording(audioBlob, "Opus"));
-            }
-
-            // Shows how to record an MP3 blob.
-            // MP3 audio is larger and cannot be not easily editible but more compatible.
-            // Also demonstrates how to use callback with PCB wrapper.
-            function recordMP3() {
-                recordCB(recorder, "mp3", (err, audioBlob) => {
-                    if (audioBlob) {
-                        addRecording(audioBlob, "MP3");
-                    }
-                });
-            }
-
-            // Show how to transcode an audio file/blob into MP3 blob using events
-            function recordOpusThenConvertToMP3() {
-                recorder.addEventListener(
-                    "recorded",
-                    (e) => {
-                        const opusBlob = e.detail.blob;
-                        recorder.convert(opusBlob, { type: "mp3" });
-                        addRecording(opusBlob, "Opus");
-                    },
-                    { once: true }
-                );
-                recorder.addEventListener(
-                    "converted",
-                    (e) => {
-                        const mp3Blob = e.detail.blob;
-                        addRecording(mp3Blob, "MP3");
-                    },
-                    { once: true }
-                );
-                recordEvents(recorder, "opus");
-            }
-
-            // Stop recording
-            function stopRecording() {
-                recorder.stop();
-            }
         </script>
         <h2>TalkRecorder Demo: iframed</h2>
         <p>Open browser console to see info on recorded blobs</p>
